requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.0
sqlmodel==0.0.16
psycopg2-binary==2.9.10
alembic==1.13.2
pydantic-settings==2.6.1
python-jose==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
PySide6==6.8.1

# Dev
pytest==8.3.3
pytest-asyncio==0.23.8
httpx==0.27.2
ruff==0.6.9
black==24.10.0
mypy==1.11.2
pre-commit==3.8.0